YARN-10005. Code improvements in MutableCSConfigurationProvider
